[Feat] Get Reqres List by Coroutine
